requirements.txt
alabaster==0.7.12
Babel==2.7.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-bootstrap-form==3.4
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-map-widgets==0.2.2
docutils==0.15.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==19.2
pkg-resources==0.0.0
Pygments==2.5.2
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
six==1.13.0
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==2.2.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
urllib3==1.25.7
